Time Order Judgement (TOJ) apparatus
A novel apparatus for biophysics experiments investigating time order
judgements. These are judgements about the perception of the order of events.

In particular, this apparatus presents pairs of LED pulses and records a
research participant's judgement of the order of the pulses and reaction time.
Temporal precision is within 1 ms.

Included here is:
- the Arduino code for the apparatus
- the Processing.org code for the apparatus controller that runs on a computer
- photos of the apparatus